Adjust sheet lookup
We used the Drive API to lookup our GSheet by name and get its GID
for handling with the Sheets API. Unfortunately, this means we need
to authenticate two different APIs with Google and as our app isn't
signed yet, we need to approve this access twice every week.

As we're always targetting the same Google Sheet anyway, let's instead
use the hard-coded sheet ID from the config JSON and skip the Drive
API call entirely.